feat: style for articles per page input + refactor css into pagination separate
