+++ src/components/pagination/pagination.css
@@ -1,0 +1,134 @@
+.pagination-container {
+    /* width: 30%; */
+    padding-top: 1rem;
+    padding-right: 10%;
+    margin-left: auto;
+    justify-content: space-between;
+    width: 30%;
+    color:  #696969;   
+}
+.pagination {
+    background-color: #addd;
+    box-shadow: 0.2rem 0.2rem #1e90ff;
+    border-radius: 0.5rem;
+    padding: 1rem;
+}
+
+.pagination a {
+  text-decoration: none;
+  color: #000000;
+  display: inline-block;
+  background: #cccccc;
+  border: 0.05rem solid #000000;
+  border-radius: 0.15rem;
+  padding: 0.25rem 0.5rem;
+  font-size: 0.7rem;
+}
+
+.pagination a:hover {
+    text-decoration: underline;
+    cursor: pointer;
+}
+
+.pagination ol {
+    list-style-type: none;
+    padding: 0;
+    margin: 0;
+    display: inline-block;
+  }
+
+.pagination  li {
+    display: inline-block;
+    margin: 0 0.125rem;
+   
+}
+
+.active a {
+    background: #ffffff;
+}
+
+.active a:hover {
+    text-decoration: none;
+    cursor: default
+}
+
+.prev.disabled, .next.disabled {
+    cursor: default;
+    background: #ccc;
+    color: #888;
+    border-color: #888;
+}
+.prev.disabled:hover, .next.disabled:hover {
+    cursor: default;
+    text-decoration: none;
+}
+
+/*hide some btns for pagination*/
+@media (min-width: 38.4rem) {
+    li:first-child, li.active-sibling, li.active, li.active + li, li:last-child {
+        display: inline-block !important;
+   }
+    li:first-child:nth-last-child(n+8) ~ li {
+        display: none;
+   }
+    li:first-child:nth-last-child(n+8) ~ li.active-sibling:before {
+        content: "\2026";
+        font-size: 1.2rem;
+        display: inline-block;
+        margin-right: 0.125rem;
+   }
+    li:first-child:nth-last-child(n+8) ~ li.active + li:after {
+        content: "\2026";
+        font-size: 1.2rem;
+        display: inline-block;
+        margin-left: 0.125rem;
+   }
+    li:first-child:nth-last-child(n+8) ~ li:nth-last-child(-n+5) {
+        display: inline-block;
+   }
+    li:first-child:nth-last-child(n+8) ~ li:nth-last-child(5):before {
+        content: "\2026";
+        font-size: 1.2rem;
+        display: inline-block;
+        margin-right: 0.125rem;
+   }
+    li:first-child:nth-last-child(n+8) ~ li:nth-child(-n+2):before, li:first-child:nth-last-child(n+8) ~ li:nth-last-child(-n+2):before, li:first-child:nth-last-child(n+8) ~ li.active-sibling:nth-last-child(-n+4):before, li:first-child:nth-last-child(n+8) ~ li:nth-child(-n+2):after, li:first-child:nth-last-child(n+8) ~ li:nth-last-child(-n+2):after, li:first-child:nth-last-child(n+8) ~ li.active-sibling:nth-last-child(-n+4):after {
+        display: none !important;
+   }
+    li:first-child:nth-last-child(n+8).active ~ li:nth-last-child(-n+5), li:first-child:nth-last-child(n+8) ~ li.active ~ li:nth-last-child(-n+5) {
+        display: none;
+   }
+    li:first-child:nth-last-child(n+8).active ~ li:nth-last-child(-n+5):before, li:first-child:nth-last-child(n+8) ~ li.active ~ li:nth-last-child(-n+5):before {
+        display: none;
+   }
+    li:first-child:nth-last-child(n+8).active ~ li:nth-child(-n+5), li:first-child:nth-last-child(n+8) ~ li.active ~ li:nth-child(-n+5) {
+        display: inline-block;
+   }
+    li:first-child:nth-last-child(n+8).active ~ li:nth-child(-n+4):after, li:first-child:nth-last-child(n+8) ~ li.active ~ li:nth-child(-n+4):after {
+        display: none;
+   }
+    li:first-child:nth-last-child(n+8).active ~ li:nth-child(5):after, li:first-child:nth-last-child(n+8) ~ li.active ~ li:nth-child(5):after {
+        content: "\2026";
+        font-size: 1.2rem;
+        display: inline-block;
+        margin-left: 0.125rem;
+   }
+    li:first-child:nth-last-child(n+8).active:before, li:first-child:nth-last-child(n+8) ~ li.active:before, li:first-child:nth-last-child(n+8).active:after, li:first-child:nth-last-child(n+8) ~ li.active:after {
+        display: none;
+   }
+}
+
+.articles-number {
+    flex-direction: column;
+    background-color: #addd;
+    box-shadow: 0.2rem 0.2rem #1e90ff;
+    border-radius: 0.5rem;
+    padding: 0.5rem;
+    font-size: 1rem;
+}
+
+.articles-number-input {
+    border-radius: 0.3rem;
+    width: 2.5rem;
+    text-align: center;
+}
