+++ src/components/pagination/pagination.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import '../../app.css';
+import './pagination.css';
 import { PaginationsProps } from '../../utils/interface';
 
 const MAX_ALLOWED_FREE_PAGES = 5; // NEWS API Regulation for free use
@@ -21,7 +21,7 @@ const Pagination: React.FC<PaginationsProps> = ({ currentPage, setCurrentPage, l
       <a onClick={() => setCurrentPage(currentPage === MAX_ALLOWED_FREE_PAGES || currentPage === pageLinksNumber()  ? currentPage : currentPage + 1)} className={currentPage === pageLinksNumber() || currentPage === MAX_ALLOWED_FREE_PAGES ? 'next disabled' : 'next'} >Next</a>
         </nav>
 
-        <span className="articles-number">
+        <span className="articles-number flex-center">
             <label htmlFor="articlesNumber">articles / page: </label>
             <input type="number" id="articlesNumber" name="articlesNumber" className="articles-number-input" min="1" max="20" defaultValue="20" />
         </span>
